+++ IPX800cs/IPX800Extensions.cs
@@ -19,7 +19,7 @@ public static class IPX800Extensions
         return ipx800.GetInputs(InputType.DigitalInput);
     }
     
-    public static int GetAnalogInput(this IIPX800 ipx800, int inputNumber)
+    public static double GetAnalogInput(this IIPX800 ipx800, int inputNumber)
     {
         return ipx800.GetAnalogInput(new AnalogInput
         {
