+++ IPX800cs/Parsers/v3/M2M/IPX800V3GetGetAnalogInputM2MResponseParser.cs
@@ -2,9 +2,9 @@
 
 internal class IPX800V3GetGetAnalogInputM2MResponseParser : IGetAnalogInputResponseParser
 {
-    public int ParseResponse(string ipxResponse, int inputNumber)
+    public double ParseResponse(string ipxResponse, int inputNumber)
     {
         ipxResponse.CheckResponse();
-        return int.Parse(ipxResponse.Trim());
+        return double.Parse(ipxResponse.Trim());
     }
 }
